+++ src/Models/MailHostname.php
@@ -1,0 +1,108 @@
+<?php
+
+namespace Cyberfusion\ClusterApi\Models;
+
+use Cyberfusion\ClusterApi\Contracts\Model;
+use Cyberfusion\ClusterApi\Support\Arr;
+
+class MailHostname extends ClusterModel implements Model
+{
+    private string $domain;
+    private int $clusterId;
+    private ?int $certificateId = null;
+    private ?int $id = null;
+    private ?string $createdAt = null;
+    private ?string $updatedAt = null;
+
+    public function getDomain(): string
+    {
+        return $this->domain;
+    }
+
+    public function setDomain(string $domain): MailHostname
+    {
+        $this->domain = $domain;
+        return $this;
+    }
+
+    public function getClusterId(): int
+    {
+        return $this->clusterId;
+    }
+
+    public function setClusterId(int $clusterId): MailHostname
+    {
+        $this->clusterId = $clusterId;
+        return $this;
+    }
+
+    public function getCertificateId(): ?int
+    {
+        return $this->certificateId;
+    }
+
+    public function setCertificateId(?int $certificateId): MailHostname
+    {
+        $this->certificateId = $certificateId;
+        return $this;
+    }
+
+    /**
+     * @return int|null
+     */
+    public function getId(): ?int
+    {
+        return $this->id;
+    }
+
+    public function setId(?int $id): MailHostname
+    {
+        $this->id = $id;
+        return $this;
+    }
+
+    public function getCreatedAt(): ?string
+    {
+        return $this->createdAt;
+    }
+
+    public function setCreatedAt(?string $createdAt): MailHostname
+    {
+        $this->createdAt = $createdAt;
+        return $this;
+    }
+
+    public function getUpdatedAt(): ?string
+    {
+        return $this->updatedAt;
+    }
+
+    public function setUpdatedAt(?string $updatedAt): MailHostname
+    {
+        $this->updatedAt = $updatedAt;
+        return $this;
+    }
+
+    public function fromArray(array $data): self
+    {
+        return $this
+            ->setDomain(Arr::get($data, 'domain'))
+            ->setClusterId(Arr::get($data, 'cluster_id'))
+            ->setCertificateId(Arr::get($data, 'certificate_id'))
+            ->setId(Arr::get($data, 'id'))
+            ->setCreatedAt(Arr::get($data, 'created_at'))
+            ->setUpdatedAt(Arr::get($data, 'updated_at'));
+    }
+
+    public function toArray(): array
+    {
+        return [
+            'domain' => $this->getDomain(),
+            'cluster_id' => $this->getClusterId(),
+            'certificate_id' => $this->getCertificateId(),
+            'id' => $this->getId(),
+            'created_at' => $this->getCreatedAt(),
+            'updated_at' => $this->getUpdatedAt(),
+        ];
+    }
+}